+++ soundlib/SampleFormatMP3.cpp
@@ -285,6 +285,17 @@ class ComponentMPG123
 #endif
 {
 	MPT_DECLARE_COMPONENT_MEMBERS
+
+private:
+
+#if MPT_OS_WINDOWS
+#if defined(MPT_WITH_MPG123)
+#elif defined(MPT_ENABLE_MPG123_DYNBIND)
+	mpt::Library MSVCRT;
+	mpt::Library LIBGCC;
+#endif // MPT_WITH_MPG123 || MPT_ENABLE_MPG123_DYNBIND
+#endif // MPT_OS_WINDOWS
+
 public:
 
 	int (*mpg123_init )(void);
@@ -355,17 +366,53 @@ public:
 		MPT_GLOBAL_BIND("mpg123", mpg123_scan);
 		MPT_GLOBAL_BIND("mpg123", mpg123_length);
 #elif defined(MPT_ENABLE_MPG123_DYNBIND)
-		AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("openmpt-mpg123")));
+		#if defined(MODPLUG_TRACKER)
+			AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("openmpt-mpg123")));
+		#endif // MODPLUG_TRACKER
+		#if MPT_OS_WINDOWS
+			// preload MSVCRT.DLL in order to prevent DLL preloading injection attacks from the current working directory,
+			// because the stock binary packages of libmpg123 link against it
+			MSVCRT = mpt::Library(mpt::LibraryPath::System(MPT_PATHSTRING("MSVCRT")));
+			#if defined(LIBOPENMPT_BUILD)
+				// require successful dependency loading for libopenmpt
+				if(!MSVCRT.IsValid())
+				{
+					return false;
+				}
+			#endif // LIBOPENMPT_BUILD
+			// preload libgcc_s_sjlj-1.dll for the same reasons (32bit only, 64bit libmpg123 does not require it)
+			MPT_CONSTANT_IF(sizeof(void*) == 4)
+			{
+				#if defined(MODPLUG_TRACKER)
+					LIBGCC = mpt::Library(mpt::LibraryPath::AppDataFullName(MPT_PATHSTRING("libgcc_s_sjlj-1"), GetComponentPath()));
+				#endif // MODPLUG_TRACKER
+				if(!LIBGCC.IsValid()) LIBGCC = mpt::Library(mpt::LibraryPath::AppFullName(MPT_PATHSTRING("libgcc_s_sjlj-1")));
+				#if defined(LIBOPENMPT_BUILD)
+					// require successful dependency loading for libopenmpt
+					if(!LIBGCC.IsValid())
+					{
+						return false;
+					}
+				#endif // LIBOPENMPT_BUILD
+			}
+		#endif // MPT_OS_WINDOWS
 		#if defined(MODPLUG_TRACKER)
 			AddLibrary("mpg123", mpt::LibraryPath::AppDataFullName(MPT_PATHSTRING("libmpg123-0"), GetComponentPath()));
 			AddLibrary("mpg123", mpt::LibraryPath::AppDataFullName(MPT_PATHSTRING("libmpg123"), GetComponentPath()));
 			AddLibrary("mpg123", mpt::LibraryPath::AppDataFullName(MPT_PATHSTRING("mpg123-0"), GetComponentPath()));
 			AddLibrary("mpg123", mpt::LibraryPath::AppDataFullName(MPT_PATHSTRING("mpg123"), GetComponentPath()));
-		#endif // MODPLUG_TRACKER
-		AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("libmpg123-0")));
-		AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("libmpg123")));
-		AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("mpg123-0")));
-		AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("mpg123")));
+			AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("libmpg123-0")));
+			AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("libmpg123")));
+			AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("mpg123-0")));
+			AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("mpg123")));
+		#elif defined(LIBOPENMPT_BUILD)
+			#if MPT_OS_WINDOWS
+				// libopenmpt on Windows only loads the official builds
+				AddLibrary("mpg123", mpt::LibraryPath::AppFullName(MPT_PATHSTRING("libmpg123-0")));
+			#else // !MPT_OS_WINDOWS
+				AddLibrary("mpg123", mpt::LibraryPath::System(MPT_PATHSTRING("mpg123")));
+			#endif // MPT_OS_WINDOWS
+		#endif
 		MPT_COMPONENT_BIND("mpg123", mpg123_init);
 		MPT_COMPONENT_BIND("mpg123", mpg123_exit);
 		MPT_COMPONENT_BIND("mpg123", mpg123_new);